requirements.txt
# AI API
openai>=1.0.0

# 数据处理
pandas>=2.0.0
numpy>=1.24.0

# 可视化
matplotlib>=3.7.0
seaborn>=0.12.0

# Web界面
gradio>=4.0.0

# 工具
python-dotenv>=1.0.0
tenacity>=8.2.0

# 测试
pytest>=7.4.0
pytest-asyncio>=0.21.0